requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
